build

+++ xv6-public/proc.c
@@ -704,14 +704,14 @@ int thread_create(thread_t *thread, void *(*start_routine)(void *), void *arg) {
   ustack[1] = (uint)arg; // argument
 
   stack -= 8;
-  if(copyout(np->pgdir, stack, ustack, 8) < 0) {
+  if(copyout(np->pgdir, (uint)stack, ustack, 8) < 0) {
     cprintf("thread_create: copyout failed!\n");
     return -1;
   }
 
   // PC, SP 설정. 여기는 exec 참고해서 다시 하기
   np->tf->eip = (uint)start_routine;
-  np->tf->esp = (uint)(stack); // argument가 하나인 경우에만 대응하나?
+  np->tf->esp = (uint)(stack); 
 
   for(i = 0; i < NOFILE; i++)
     if(curproc->ofile[i])
